Minor changes in Diablo, Rogue, Warrior, Wizard.
Sort of phased out the player class, depending on how we do the XML stuff this may change. The way I was thinking to do the XML was just put all of the possible methods we could use for bosses/players(abilities, change healths, names, etc.) and then each specific class would have their own call to the XML parser. It makes more sense if I actually explain it in person.

GameClient and GameServer now exclusively use object streams for all network traffic. I know we were initially planning to use multiple streams but this is a bit cleaner and I was running into issues with opening multiple I/O streams on each socket anyway. As of now, when the fighters connect they get added to a vector of Fighters including the boss. I'm still working on sending out the updated vector everytime something combat related happens, but that shouldn't be too bad.

tl;dr Chat still works as it did before, but now we're tossing around some fancy objects.