Took out the main method of GameClient so now you need to start the game from the Main Menu, which is how it will be in the end. Feel free to add a main for testing as needed.
MainMenu now has some error checking on the IP address/port number. If there's no port number or IP address, it will prompt for one. If there is an ip address/port it will try to connect; if the connection is successful it will start a new game provided that a fighter has been selected. If the connection is unsuccessful, it will say so.

As of now the Main Menu just needs a GUI component added to show the currently selected fighter, it's being stored as myFighter in the class itself, it just needs to be assigned to some JLabel or something.