Updated the icons for the players to 50x50, makes it a bit easier to work with.
Added in getIcon methods for Fighter, Warrior, Wizard, Rogue, and GameDataParser. All of the "heavy lifting" of reading in the icon and casting it as an ImageIcon is done in GameDataParser. The way that images loaded through the getIcon() method should make it so Zihao can see the pictures too since it's path independent.

Changed around the SelectionScreen and got it properly wrapping words. Well sorta, the height and width of the frame need to be hard coded and then the text will wrap. For now everything is working fine as long as we don't have any super long descriptions.

Changed MainMenu a little bit to require a class to be selected before the game will start. A JLabel or something should be added somewhere to the menu to indicate which class is currently selected. I didn't want to mess with that stuff too much since Zihao was working on cleaning up some GUIs I think.

Changed the GameClient to start pulling some information from the fighters, only did it for the player that you first connect with for now mostly just for testing. Currently, it pulls the icon, name, class, and base/current health. In the end all the information will need to update as players connect so it will have to be pulled from the clientList Vector.